[DEV] Moved Code.java into the code package.
