+++ src/main/java/controllers/CodeViewControllers/CodeCellController.java
@@ -9,7 +9,7 @@ import javafx.scene.input.ClipboardContent;
 import javafx.scene.input.KeyCode;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.*;
-import models.Code;
+import code.readers.Code;
 
 public class CodeCellController {
     @FXML
